[SID:3376] PO 실 결함 + 유나 design verdict 2건 반영
**슬러그 충돌(PO 실측, 서버 기동 자체 실패)** — channel-connections 아래
[channel](authorize·callback·app-credentials)과 [connectionId](disconnect·
test)가 같은 깊이에 다른 동적 세그먼트 이름이라 Next.js가 라우팅 트리를
못 만들어 `next dev`/`next build` 후 서버 기동이 죽는 급이었다. disconnect·
test를 [channel] 폴더 아래로 통합(폴더명 하나로 통일 — Next 제약)하고,
라우트 내부에서 `channel: connectionId`로 되짚어 원래 의미·BE 경로 템플릿은
그대로 보존. `pnpm dev` 실 기동으로 재확認(Ready in 316ms, 에러 0).

**design verdict 2건**:
① app-credentials-card.tsx — effective_source==='none'(설정 미완)을
destructive/alert로 그리던 것을 default/status로(같은 화면 not_connected
칩도 중립 — "아직 안 한 것"이지 실패가 아니다).
② channels/page.tsx — 연결 시험 결과 텍스트의 계열색 직접 사용(라이트
성공 3.49, 4.5 미달)을 ChannelStatusChip과 같은 원칙(dot으로 성패·텍스트는
text-foreground)으로.

vitest 5403·eslint·tsc·build 전부 클린.
